move _validate_inputs into its own module
so that other modules can import it
